refactor: reorganize component imports and create new TextInput and PasswordInput components
