+++ src/views/Shared/index.ts
@@ -1,0 +1,1 @@
+export { default as SharedViewWrapper } from './SharedViewWrapper';
