--- src/views/Shared/index.tsx
@@ -1,6 +1,0 @@
-export { default as SharedViewWrapper } from './SharedViewWrapper';
-export { default as SharedView } from './SharedView';
-export * from './components/SharedItemList';
-export * from './components/ShareItemDialog/ShareItemDialog';
-export * from './hooks/useFetchSharedData';
-export * from './hooks/useSharedContextMenu';
